+++ Bugsnag/include/Bugsnag/Bugsnag.h
@@ -289,6 +289,8 @@
  *  @param userId ID of the user
  *  @param name   Name of the user
  *  @param email  Email address of the user
+ *
+ *  If user ID is nil, a Bugsnag-generated Device ID is used for the `user.id` property of events and sessions.
  */
 + (void)setUser:(NSString *_Nullable)userId
        withEmail:(NSString *_Nullable)email