+++ Bugsnag/KSCrash/Source/KSCrash/Recording/Tools/BSG_KSMach.c
@@ -224,8 +224,8 @@ pthread_t bsg_ksmachpthreadFromMachThread(const thread_t thread) {
 
 bool bsg_ksmachgetThreadName(const thread_t thread, char *const buffer,
                              size_t bufLength) {
-    // WARNING: This implementation is only async-safe for the current thread
-    // as of libpthread-330.201.1, and is still unsafe for other threads.
+    // WARNING: This implementation is not async-safe because
+    // pthread_from_mach_thread_np() acquires an internal lock.
 
     const pthread_t pthread = pthread_from_mach_thread_np(thread);
     if (pthread == NULL) {